requirements.txt
# Telegram Mental Health Bot - Dependencies
# Python 3.8+ required

# Core Bot Framework
python-telegram-bot==20.7    # Telegram Bot API wrapper

# Environment & Configuration
python-dotenv==1.0.0         # Environment variable management from .env files

# Scheduling & Time Management
apscheduler==3.10.4          # Advanced Python Scheduler for reminders
pytz==2023.3                 # Timezone definitions for worldwide support
python-dateutil==2.8.2       # Extensions to standard datetime module

# AI Integration
openai==1.12.0               # OpenAI/OpenRouter API client for weekly AI reports

# Optional: Development Dependencies (uncomment if needed)
# pytest==7.4.3              # Testing framework
# black==23.12.0             # Code formatter
# pylint==3.0.3              # Code linter
# mypy==1.7.1                # Static type checker
